Separated stats writer part from apputil
