+++ apps/support.maf
@@ -9,6 +9,7 @@
 
 SOURCES
 apputil.cpp
+statswriter.cpp
 logsupport.cpp
 logsupport_appdefs.cpp
 socketoptions.cpp
