+++ testing/srt-test-live.maf
@@ -4,6 +4,7 @@ SOURCES
 srt-test-live.cpp
 testmedia.cpp
 ../apps/apputil.cpp
+../apps/statswriter.cpp
 ../apps/verbose.cpp
 ../apps/socketoptions.cpp
 ../apps/uriparser.cpp
